Add files via upload
We present you the 10 athletes with the highest incomes in 2020-2021. You will know their age range, in which part of the world they're come from and what kind of sports they practice. The goal is to show and understand how those qualitative variables interact with one another over the income and how taken separately they influence the incomes of those players. The results will be analyzed and shown through regression analysis, statistics and graphics.